backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.35.0
torch==2.1.0
yfinance==0.2.32
pandas==2.1.3
python-dotenv==1.0.0
groq==0.4.1
requests==2.31.0